First commited as before_15104_rev1

+++ lttng/org.eclipse.linuxtools.ctf.core.tests/src/org/eclipse/linuxtools/ctf/core/tests/io/BitBufferTest.java
@@ -143,7 +143,7 @@ public class BitBufferTest {
      */
     @Test
     public void testGetPosition() {
-        long result = fixture.position();
+        int result = fixture.position();
 
         assertEquals(1, result);
     }
